Renaming directory.... VS Code did some refactoring
