+++ adaptivetesting/_prompt_builder.py
@@ -2,7 +2,7 @@ import numpy as np
 import logging
 import re
 import urllib.parse
-import adatest
+import adaptivetesting
 from .embedders import cos_sim
 from .utils import is_subtopic
 log = logging.getLogger(__name__)
@@ -158,13 +158,13 @@ class PromptBuilder():
             if self.prompt_diversity:
                 sim_avoidance = np.zeros(len(ids))
                 if suggest_topics:
-                    embeddings_arr = np.vstack(adatest.embed(
+                    embeddings_arr = np.vstack(adaptivetesting.embed(
                         [urllib.parse.unquote(test_tree.loc[id, "topic"].split("/")[-1]) for id in ids]
                     ))
                 else:
                     embeddings_arr = np.hstack([
-                        np.vstack(adatest.embed([test_tree.loc[id, "input"] for id in ids])),
-                        np.vstack(adatest.embed([test_tree.loc[id, "output"] for id in ids]))
+                        np.vstack(adaptivetesting.embed([test_tree.loc[id, "input"] for id in ids])),
+                        np.vstack(adaptivetesting.embed([test_tree.loc[id, "output"] for id in ids]))
                     ])
                 similarities = cos_sim(embeddings_arr, embeddings_arr)
             hard_avoidance = np.zeros(len(ids))
