+++ adaptivetesting/resources/main.js
