+++ adaptivetesting/resources/main.js.LICENSE.txt
